mcp-server/requirements.txt
fastmcp
starlette
mcp
requests
elevenlabs
python-dotenv
Pillow
numpy
shotstack-sdk
srt
certifi
PyJWT
ffmpeg-python